home: Rework the mechanism for generating graphs via custom template
Previously using python-nvd3 we were constrained by the templates
distributed and the properties exposed by the python
bindings/abstraction for nv d3 which were incomplete. To avoid multiple
passes on the page (sever side and client side) we can just generate the
graph with the correct js/html in the first place.

Signed-off-by: Michael Wood <michael.g.wood@intel.com>